+++ include/SyncedSchema.h
@@ -52,8 +52,8 @@ public:
   std::optional<double> exclusiveMinimum_;
 
   // String properties
-  std::optional<double> maxLength_;
-  std::optional<double> minLength_;
+  std::optional<size_t> maxLength_;
+  std::optional<size_t> minLength_;
   std::optional<std::string> pattern_;
 
   // Array properties
